Display card in row
