+++ src/screens/home/retaliation/HomeRetaliationRowContainer.tsx
@@ -1,13 +1,10 @@
 import React, { FunctionComponent } from 'react'
-import { StyleSheet, Text, View, Image } from 'react-native'
-import { Spacing, Styles, Typography } from '../../../styles'
-import CardView from '../../shared/CardView'
 import { HomeRetaliationRowContainerViewModel } from '../HomeRowViewModel'
-import { useTheme, useThemedStyles } from '../../../themes'
-import { BorderlessButton, PrimaryButton } from '../../shared/Buttons'
-import Theme from '../../../themes/Theme'
-import { HorizontalSeparator } from '../../shared/HorizontalSeparator'
-import i18n from '../../../utils/i18n'
+import { ScrollView } from 'react-native-gesture-handler'
+import RetaliationCard from './RetaliationCard'
+import { RetaliationCardViewModel } from './RetaliationCardViewModel'
+import { StyleSheet, View } from 'react-native'
+import { Spacing } from '../../../styles'
 
 type Props = Readonly<{
   viewModel: HomeRetaliationRowContainerViewModel
@@ -16,73 +13,32 @@ type Props = Readonly<{
 const HomeRetaliationRowContainer: FunctionComponent<Props> = ({
   viewModel,
 }) => {
-  const { theme } = useTheme()
-  const styles = useThemedStyles(stylesFactory)
   return (
-    <CardView style={styles.cardView} backgroundColor={theme.lightBackground}>
-      <View style={styles.container}>
-        <Text>
-          <Image source={require('../../../assets/images/iconSearch.png')} />
-          <Text style={styles.title}>{viewModel.title}</Text>
-        </Text>
-        <Text style={styles.body} numberOfLines={4}>
-          {viewModel.body}
-        </Text>
-        <BorderlessButton
-          title={i18n.t('home.retaliation.see_more')}
-          textStyle={styles.linkText}
-          style={styles.linkButton}
-          onPress={() => {}}
-        />
-        <HorizontalSeparator />
-        <PrimaryButton
-          shape={'rounded'}
-          disabled={false}
-          style={styles.retaliateButton}
-          textStyle={styles.retaliateButtonText}
-          title={i18n.t('home.retaliation.retaliate_button')}
-          onPress={() => {}}
-        />
-      </View>
-    </CardView>
+    <ScrollView>
+      <Cards viewModel={viewModel} />
+    </ScrollView>
   )
 }
 
-const stylesFactory = (theme: Theme) => {
+const Cards: FunctionComponent<Props> = ({ viewModel }) => {
+  const styles = stylesFactory()
+  return (
+    <View style={styles.container}>
+      {viewModel.retaliations.map(
+        (retaliationViewModel: RetaliationCardViewModel) => {
+          return <RetaliationCard viewModel={retaliationViewModel} />
+        },
+      )}
+    </View>
+  )
+}
+
+const stylesFactory = () => {
   return StyleSheet.create({
-    body: {
-      ...Typography.body,
-      marginTop: Spacing.margin,
-    },
-    cardView: {
-      marginHorizontal: Spacing.margin,
-      marginVertical: Spacing.margin,
-    },
     container: {
       padding: Spacing.margin,
-    },
-    linkButton: {
-      alignSelf: 'flex-start',
-      textAlign: 'left',
-      paddingHorizontal: 0,
-      paddingVertical: 12,
-    },
-    linkText: {
-      ...Styles.eventSeeMoreButtonTextStyle(theme),
-      ...Typography.body,
-      color: theme.primaryColor,
-    },
-    retaliateButton: {
-      marginTop: Spacing.mediumMargin,
-    },
-    retaliateButtonText: {
-      ...Typography.headline,
-      color: theme.primaryButtonTextColor,
-    },
-    title: {
-      ...Typography.title2,
-      lineHeight: 21,
-      marginBottom: Spacing.margin,
+      flexDirection: 'row',
+      alignItems: 'center',
     },
   })
 }
