+++ src/screens/home/retaliation/RetaliationCardViewModel.tsx
@@ -1,0 +1,5 @@
+export interface RetaliationCardViewModel {
+  id: string
+  title: string
+  body: string
+}
